hub.c: Don't use VLA in nmdc_search() to fix segfault on aarch64
Fixes #102. Or rather, works around it. I don't really know yet *why*
this is causing a segfault on aarch64 - it could be a compilation bug or
an actual bug in ncdc. In the latter case, this workaround could be
masking a more serious issue, but alas, so far all debugging output
seems to point to stack corruption, we've not managed to reproduce it on
other arches and this doesn't seem easy to investigate futher. :(